Refactor localStorage and fix bug
